fix: upload avatar logic

+++ src/services/userService.ts
@@ -1,4 +1,4 @@
-import { User } from "@/models";
+import { Post, User } from "@/models";
 import { Model, ModelStatic, UniqueConstraintError } from "sequelize";
 import { AbstractService } from "./abstractService";
 import { Request } from "express";
@@ -12,7 +12,9 @@ export class UserService extends AbstractService<User> {
     }
 
     findAll = () => {
-        return User.findAll();
+        return User.findAll({
+            include: { model: Post, as: 'posts'},
+        });
     };
     findById = (id: number) => {
         return User.findOne({ where: { id } });
@@ -24,14 +26,18 @@ export class UserService extends AbstractService<User> {
     }
     update = async (id: number, data: User): Promise<[affectedCount: number]> => {
         const user = await this.findById(id);
+        
         if(user) {
-            const avatar = user.avatar;
-            Files.removePublicFileSyncByPath(avatar);
+            if (data.dataValues.hasOwnProperty('avatar')) {
+                const avatar = user.avatar;
+                Files.removePublicFileSyncByPath(avatar);
+            }
             delete data.dataValues.id;
             return await User.update({ ...data.dataValues }, { where: { id } });            
         }
         throw new UserNotFoundError();
     };
+
     delete = async (id: number): Promise<number> => {
         const user = await this.findById(id);
         if(user) {
@@ -46,12 +52,19 @@ export class UserService extends AbstractService<User> {
         const file = req.file;
         const avatar = Files.getPublicPath(file);
         const hashedPassword = Decrypt.sha256(req.body.password);
-        const user = User.build({ ...req.body, password: hashedPassword, DOB: new Date(req.body.DOB), avatar });
+        let user: User;
+        if(req.body.DOB)
+            user = User.build({ ...req.body, password: hashedPassword, DOB: new Date(req.body.DOB), avatar });
+        else user = User.build({ ...req.body, password: hashedPassword, avatar });
         try {
             if(isCreate) {
                 const createdUser = await this.create(user);
                 return createdUser;
             } else {            
+                const { isUpload } = req.body;
+                if(!avatar && !isUpload || isUpload === 'false')
+                    delete user.dataValues.avatar;
+                
                 const id = parseInt(req.params.id);
                 return await this.update(id, user);
             }
